+++ src/transaction/mod.rs
@@ -1,4 +1,6 @@
 pub mod endorsed;
+pub mod executor;
 pub mod mvcc;
+pub mod parallel;
 pub mod proposal;
 pub mod rwset;